requirements.txt
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==4.4.2
easyocr==1.7.1
executing==2.0.1
ffmpeg==1.4
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
fsspec==2023.12.2
google==3.0.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-language==2.12.0
google-cloud-speech==2.23.0
google-cloud-storage==2.13.0
google-cloud-texttospeech==2.15.0
google-cloud-translate==3.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
google-speech==1.2.0
google-translate==0.1
googleapis-common-protos==1.62.0
googletrans==3.1.0a0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.33.1
imageio-ffmpeg==0.4.9
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lazy_loader==0.3
libretranslatepy==2.1.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
proglog==0.1.10
prompt-toolkit==3.0.41
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pvleopard==2.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydub==0.24.1
Pygments==2.17.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
termcolor==2.4.0
terminado==0.18.0
testpath==0.6.0
tifffile==2023.12.9
toml==0.10.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
translate==3.6.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
web-cache==1.1.0
webencodings==0.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
